% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tag.R
\name{add_tag}
\alias{add_tag}
\title{Add HTML Tags to Text}
\usage{
add_tag(text, tag = "h3")
}
\arguments{
\item{text}{Text string}

\item{tag}{HTML tag like p, b, h1 or other}
}
\value{
\code{glue} \code{character} with HTML tags
}
\description{
Add HTML Tags to Text
}
\examples{
add_tag("Title Text", tag = "h2")
add_tag("bold text", tag = "b")
}
\seealso{
You can add tags to text using \code{\link[=clean_tags]{clean_tags()}} function
}
