% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tags.R
\name{clean_tags}
\alias{clean_tags}
\title{Clean HTML Tags from the Text}
\usage{
clean_tags(html_string)
}
\arguments{
\item{html_string}{Text with html tags}
}
\value{
\code{\link[=character]{character()}} vector with text strings without tags
}
\description{
Clean HTML Tags from the Text
}
\examples{
clean_tags("<b>Total increase is equal to 14.5 \% </b>")
clean_tags("<h2>Sales by Region</h3>")
}
\seealso{
You can add tags to text using \code{\link[=add_tag]{add_tag()}} function
}
