% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text String}
\usage{
clean_text(
  text,
  upper = c("YTD", "PYTD"),
  lower = c("vs", "br>", "h1>", "h2>", "h3>", "h4>", "h5>", "h6>", "b>", "<span",
    "span>")
)
}
\arguments{
\item{text}{Text string for cleaning. Can contain multiple sentences.}

\item{upper}{Vector of words that need to be changed to uppercase in text}

\item{lower}{Vector of words that need to be changed to lowercase in text}
}
\value{
Text string
}
\description{
Function cleans up the text string improving narration, removing excessive white spaces
}
\examples{
text <- "Similarly in 2020 the sum of spend increased by 15.4\%  ( 4.3 \% higher  than average)."
clean_text(text)
clean_text(" Total  is 12,300 Orders ( 23.5 \% for East  ) ")
}
