% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_text.R
\name{correct_text}
\alias{correct_text}
\title{Correct the sentence using udpipe}
\usage{
correct_text(text, language = "english", download_udpipe = TRUE, measure = "")
}
\arguments{
\item{text}{Text string for correction}

\item{language}{Language for udpipe model, defaults to 'english'}

\item{download_udpipe}{Download udpipe to working directory}

\item{measure}{Measure for the template}
}
\value{
Text string
}
\description{
Correct the sentence using udpipe
}
\examples{
correct_text("In 2020 total Sales across all Regions (EMEA, NA, ASPAC) is equal to 23.5 M")
correct_text("Total Profit across all Regions are 85 M")
}
