% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Format numeric value}
\usage{
format_num(
  num,
  format = "auto",
  decimals = 1,
  threshold = 10000,
  format_list = c("", "K", "M", "B", "T")
)
}
\arguments{
\item{num}{Number or numeric vector}

\item{format}{Format for numeric output - "auto", "K", "M", "B" or "T"}

\item{decimals}{Number of decimal numbers for rounding}

\item{threshold}{Threshold for converting to a specific format, below threshold only thousand delimiters will be added}

\item{format_list}{List of formats to use for automated}
}
\value{
vector of numerics
}
\description{
Format numeric value
}
\examples{
format_num(27334254)
format_num(c(27334254, 12332, 23425435534))
}
