% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_text.R
\name{format_text}
\alias{format_text}
\title{Add HTML tags to colorize and bold the string of text}
\usage{
format_text(text, color = "auto", bold = TRUE)
}
\arguments{
\item{text}{Text string that you want to format}

\item{color}{Color name, for "auto" the color will be determined based on the number parsed - red for negative, green for positive}

\item{bold}{Make text bold or not}
}
\value{
text with HTML tags
}
\description{
Function adds html code around the selected string adding options for colorized and/or bold display in HTML documents.
}
\examples{
text <- format_text("1.2\%", color = "auto", bold = TRUE)
text
}
