% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequency.R
\name{get_frequency}
\alias{get_frequency}
\title{Get Date Frequency of the Data}
\usage{
get_frequency(df, date_field = NULL)
}
\arguments{
\item{df}{\code{\link[=data.frame]{data.frame()}} or \code{\link[=tibble]{tibble()}} Data frame of tibble, can be aggregated or raw}

\item{date_field}{Date field to be analyzed, by default the first date-like column will be used}
}
\value{
frequency - "quarter", "month", "week" or "day"
}
\description{
Function will group the data and estimate the frequency of the time stamps,
returning 'year', 'quarter', 'month', 'week' or 'day'. You can use it on raw or aggregated data frames
}
\examples{
sales \%>\%
  dplyr::mutate(Date = lubridate::floor_date(Date, unit = "month")) \%>\%
  get_frequency()

sales \%>\%
  dplyr::mutate(Date = lubridate::floor_date(Date, unit = "week")) \%>\%
  get_frequency()

sales \%>\%
  dplyr::mutate(Date = lubridate::floor_date(Date, unit = "quarter")) \%>\%
  dplyr::group_by(Region, Date) \%>\%
  dplyr::summarise(Sales = sum(Sales, na.rm = TRUE)) \%>\%
  get_frequency()

get_frequency(sales)
}
