% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{narrate_desc}
\alias{narrate_desc}
\title{Create Descriptive Narrative}
\usage{
narrate_desc(
  df,
  measure = NULL,
  dimensions = NULL,
  coverage = 0.5,
  coverage_limit = 5,
  narrative_total = "{measure} across all {pluralize(dimension1)} is {total}. ",
  narrative_outlier = "Outlying {dimension} by {measure} is {outlier_insight}. ",
  narrative_outlier_multiple =
    "Outlying {pluralize(dimension)} by {measure} are {outlier_insight}. ",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Create Descriptive Narrative
}
