% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{narratorServer}
\alias{narratorServer}
\title{Server Logic for Narrator}
\usage{
narratorServer(
  id,
  df,
  narrative_type = "descriptive",
  summarization = "sum",
  ...
)
}
\arguments{
\item{id}{Namespace specification}

\item{df}{Reactive data frame}

\item{narrative_type}{Type of narrative 'descriptive', 'trend' or 'forecast'}

\item{summarization}{Summarization/aggregation for the data - 'sum', 'count' or 'average'}

\item{...}{Additional arguments for \verb{narrate_*} functions}
}
\value{
Narrative output
}
\description{
Server logic for the shiny module to work together with \link{narratorUI}
}
\examples{
library(shiny)
library(narrator)
library(dplyr)

df <- sales \%>\%
  group_by(Region, Product) \%>\%
  summarise(Sales = sum(Sales, na.rm = TRUE))

ui <- fluidPage(
  narratorUI(id = "main")
)

server <- function(input, output, session) {
  narratorServer(
    id = "main",
    df = reactive(df),
    format = TRUE,
    summarization = "sum")
}

if (interactive()) shinyApp(ui, server)
}
