% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module.R
\name{narratorUI}
\alias{narratorUI}
\title{Shiny Module for Narrator}
\usage{
narratorUI(
  id,
  narration_depth = 2,
  coverage = 0.5,
  coverage_limit = 5,
  format = TRUE,
  ...
)
}
\arguments{
\item{id}{Namespace specification}

\item{narration_depth}{Parameter to control the depth of the analysis 1 for summary and 2 for detailed}

\item{coverage}{Numeric portion of variability to be covered by narrative, 0 to 1}

\item{coverage_limit}{Integer maximum number of elements to be narrated, overrides
coverage to avoid extremely verbose narrative creation}

\item{format}{If \code{TRUE} - format text with HTML tags}

\item{...}{other arguments passed to \code{\link[glue]{glue}}}
}
\value{
UI
}
\description{
UI for the shiny module to work together with \link{narratorServer}
}
\examples{
library(shiny)
library(narrator)
library(dplyr)

df <- sales \%>\%
  group_by(Region, Product) \%>\%
  summarise(Sales = sum(Sales, na.rm = TRUE))

ui <- fluidPage(
  narratorUI(id = "main")
)

server <- function(input, output, session) {
  narratorServer(
    id = "main",
    df = reactive(df),
    format = TRUE,
    summarization = "sum")
}

if (interactive()) shinyApp(ui, server)
}
