% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pluralize}
\alias{pluralize}
\title{Pluralize a word}
\usage{
pluralize(x, n = 2, prepend = FALSE)
}
\arguments{
\item{x}{character vector of words to make plural}

\item{n}{amount of plural (some plurals change by amount and this also impacts
the \code{prepend} functionality)}

\item{prepend}{should we prepend the value of \code{n} to the output? Default: \code{FALSE}}
}
\value{
character vector of modified pluralized words
}
\description{
Pluralize a word
}
\examples{
pluralize('test')
pluralize('test', 5)
pluralize('test', 5, TRUE)
pluralize('regex')
}
