% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Synthetic Sales Data}
\format{
A tibble with 10000 rows and 9 variables:
\describe{
\item{Order ID}{Unique ID for the order}
\item{Date}{Date of Order creation}
\item{Region}{Region/Region of shipment}
\item{Product}{Product Name}
\item{Store}{Store size - Big, Medium or Small}
\item{Promotion}{Was Promotion present}
\item{Price}{Raw unit price for quantity}
\item{Quantity}{Number of units sold}
\item{Sales}{Final Sales, including promotional discounts}
}
}
\usage{
sales
}
\description{
Synthetic Sales Data
}
\examples{
sales
}
\keyword{datasets}
