% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_html.R
\name{to_html}
\alias{to_html}
\title{Format the Narrative with HTML tags for Interactive Reports and Applications}
\usage{
to_html(narrative, format_percentage = TRUE, title_tag = "h4", ...)
}
\arguments{
\item{narrative}{List of narratives returned by narrate_* function}

\item{format_percentage}{Add percentage formatting to text}

\item{title_tag}{HTML tag to be added to the names of the narrative}

\item{...}{Other arguments passed to \code{format_pct()} function}
}
\value{
html \code{\link[=character]{character()}} of text strings
}
\description{
Format the Narrative with HTML tags for Interactive Reports and Applications
}
\examples{
library(dplyr)

sales \%>\%
group_by(Region, Product) \%>\%
  summarise(Sales = sum(Sales, na.rm = TRUE)) \%>\%
  arrange(desc(Sales)) \%>\%
  narrate_descriptive() \%>\%
  to_html()
}
