% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt.R
\name{translate_narrative}
\alias{translate_narrative}
\title{Translate the narrative output with ChatGPT}
\usage{
translate_narrative(
  narrative,
  language,
  openai_api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{narrative}{List of narratives returned by narrate_* function, character vector or string that will be enhanced by ChatGPT}

\item{language}{Write a language to translate narrative to, please use English here, ex. "Spanish" not "Español"}

\item{openai_api_key}{Your OpenAI API key, you can set it up in .Renviron file as "OPENAI_API_KEY", function will look for it with \code{Sys.getenv("OPENAI_API_KEY")}}
}
\value{
\code{\link[=character]{character()}} of narratives enhanced by ChatGPT
}
\description{
Translate the narrative output with ChatGPT
}
\examples{
\dontrun{
narrative <- sales \%>\%
dplyr::filter(Product \%in\% c("Tools", "Clothing", "Home")) \%>\%
  dplyr::group_by(Product, Region)  \%>\%
  dplyr::summarise(Sales = sum(Sales)) \%>\%
  narrate_descriptive()

narrative <- enhance_narrative(narrative)
translate_narrative(narrative, "Spanish")
}
}
