% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tag.R
\name{add_tag}
\alias{add_tag}
\title{Add HTML tags to text}
\usage{
add_tag(text, tag = "h3")
}
\arguments{
\item{text}{Text string}

\item{tag}{HTML tag like p, b, h1 or other}
}
\value{
Text string with HTML tags
}
\description{
Add HTML tags to text
}
\examples{
add_tag("Title Text", tag = "h2")
add_tag("bold text", tag = "b")
}
