% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tags.R
\name{clean_tags}
\alias{clean_tags}
\title{Clean html tags from the text}
\usage{
clean_tags(html_string)
}
\arguments{
\item{html_string}{Text with html tags}
}
\value{
Text string without tags
}
\description{
Clean html tags from the text
}
\examples{
clean_tags("<b class = ttip>Total increase is equal to 14.5 \% <span class='tooltiptext'> from 1300 to 1670</span> </b>")
}
