% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_templates.R
\name{list_templates}
\alias{list_templates}
\title{Return a Table of all Templates Currently Available in the Package}
\usage{
list_templates()
}
\value{
a tibble with all templates and functions in the packages
}
\description{
Return a Table of all Templates Currently Available in the Package
}
\examples{
list_templates()
}
