% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrate_descriptive.R
\name{narrate_descriptive}
\alias{narrate_descriptive}
\title{Create Descriptive Narrative}
\usage{
narrate_descriptive(
  df,
  measure = NULL,
  dimensions = NULL,
  summarization = "sum",
  coverage = 0.5,
  coverage_limit = 5,
  narration_depth = 2,
  template_total = "Total {measure} across all {pluralize(dimension1)}: {total}.",
  template_outlier = "Outlying {dimension} by {measure} is {outlier_insight}.",
  template_outlier_multiple =
    "Outlying {pluralize(dimension)} by {measure} are {outlier_insight}.",
  template_outlier_l2 =
    "In {root_outlier_dimension}, significant {dimension_l2}\\n    by {measure} is {outlier_insight}.",
  template_outlier_l2_multiple =
    "In {root_outlier_dimension}, significant {pluralize(dimension_l2)}\\n    by {measure} are {outlier_insight}.",
  use_renviron = FALSE,
  return_data = FALSE,
  simplify = FALSE,
  format_numbers = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame of tibble, can be aggregated or raw}

\item{measure}{Numeric measure for function to create calculations with,
if NULL then it will take the first numeric field available}

\item{dimensions}{Vector of dimensions for analysis, by default all character
or factor variable will be used}

\item{summarization}{Approach for data summarization/aggregation - 'sum', 'count' or 'mean'}

\item{coverage}{Portion of variability to be covered by narrative, 0 to 1}

\item{coverage_limit}{Maximum number of elements to be narrated, overrides
coverage to avoid extremely verbose narrative creation}

\item{narration_depth}{Parameter to control the depth of the analysis 1 for summary and 2 for detailed}

\item{template_total}{glue template for total volumes narrative}

\item{template_outlier}{glue template for single outlier narrative}

\item{template_outlier_multiple}{glue template for multiple outliers narrative}

\item{template_outlier_l2}{glue template for deeper hierarchical single outlier narrative}

\item{template_outlier_l2_multiple}{glue template for deeper hierarchical multiple outliers narrative}

\item{use_renviron}{Whether to use .Renviron variables in the template}

\item{return_data}{return a list of variables used in the function's templates}

\item{simplify}{if TRUE - returns a character vector, if FALSE - named list}

\item{format_numbers}{Whether to format big numbers to K/M/B using format_num function}

\item{...}{other arguments passed to glue::glue}
}
\value{
character vector, glue
}
\description{
Create Descriptive Narrative
}
\examples{
narrator::sales \%>\%
narrate_descriptive(measure = "Sales",
            dimensions = c("Region", "Product"))

narrator::sales \%>\%
  dplyr::filter(Product \%in\% c("Product A", "Product B")) \%>\%
  dplyr::group_by(Product, Region)  \%>\%
  dplyr::summarise(Quantity = sum(Quantity)) \%>\%
  narrate_descriptive()
}
