% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrate_descriptive.R
\name{narrate_descriptive}
\alias{narrate_descriptive}
\title{Create Descriptive Narrative}
\usage{
narrate_descriptive(
  df,
  measure = NULL,
  dimensions = NULL,
  coverage = 0.5,
  coverage_limit = 5,
  template_total = "{measure} across all {pluralize(dimension1)} is {total}. ",
  template_outlier = "Outlying {dimension} by {measure} is {outlier_insight}. ",
  template_outlier_multiple =
    "Outlying {pluralize(dimension)} by {measure} are {outlier_insight}. ",
  use_renviron = FALSE,
  return_data = FALSE,
  ...
)
}
\arguments{
\item{df}{Data frame of tibble, can be aggregated or raw}

\item{measure}{Numeric measure for function to create calculations with,
if NULL then it will take the first numeric field available}

\item{dimensions}{Vector of dimensions for analysis, by default all character
or factor variable will be used}

\item{coverage}{Portion of variability to be covered by narrative, 0 to 1}

\item{coverage_limit}{Maximum number of elements to be narrated, overrides
coverage to avoid extremely verbose narrative creation}

\item{template_total}{glue template for total volumes narrative}

\item{template_outlier}{glue template for single outlier narrative}

\item{template_outlier_multiple}{glue template for multiple outliers narrative}

\item{use_renviron}{Whether to use .Renviron variables in the template}

\item{return_data}{return a list of variables used in the function's templates}

\item{...}{other arguments passed to glue::glue}
}
\value{
character vector, glue
}
\description{
Create Descriptive Narrative
}
\examples{
sales \%>\%
narrate_descriptive(measure = "Sales",
            dimensions = c("Territory", "Product"))
}
